+++ src/Infrastructure/Console/ElasticDelete.php
@@ -19,6 +19,8 @@ class ElasticDelete extends Command
 
     public function handle(ElasticIndexAdapter $adapter, IndexConfigurationRepositoryInterface $indexConfigurationRepository): int
     {
+        $this->warn('This command is deprecated and will be removed in 3.0. Use scout:delete-index `name` instead.');
+
         $config = config('explorer');
 
         if (!$config) {