+++ changelog.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [2.5.0]
+
 ### Added
 - The new match_phrase query (#73)
 - Ability to optionally set more parameters for a few queries (#68)